store.txt, moagyBot.py, nltkCorpus.txt, nltkTestFuck.py
Added more lists to the store file.
Added specifically fitness related corpus to the nltkCorpus file.
Modified the moagyBot file and added comments so you guys can read and hopefully get the code, also referenced the websites I used to make it all and the guides I followed.
The nltkTestFuck file is a better moagyBot file pretty much. Tbh the moagyBot file is pretty shit code and most of it doesn't make sense and isn't commented on.